CLIENTS-1476: Create Kafka HTTP Server interfaces and loading mechanism. (#1628)
This PR introduces an interface for a HTTP server to be run inside the broker, and a mechanism to load such server from the classpath (to avoid cyclic dependencies between repositories). Binding of the actual implementation of the Kafka HTTP Server is done by registered a implementation of KafkaHttpServerProvider in the classpath (see KafkaHttpServerLoader for details about loading).

KafkaHttpServer has a life-cycle detached from the broker life-cycle. The server is started after the broker starts up, and is stopped before the broker shuts down.

Runtime dependencies are passed down from the broker to the server at loading time via KafkaHttpServerInjector. A included KafkaHttpServerBinder can be used to construct such injector.

This new Kafka HTTP Server is going to be used to substitute MetadataService as a means of running an HTTP server in the broker (to completely separate it from RBAC). Changes in ConfluentProvider, AuthorizerServerInfo, and Broker.scala are to add the RBAC dependencies in the KafkaHttpServerBinder, so that RBAC (HTTP service) can be run in the new HTTP server, and MetadataService can be deleted.

Co-authored-by: Stanislav Kozlovski <stanislav_kozlovski@outlook.com>