CNKAF-638: Introduce SBK internal RPC client (#1644)
Part of https://confluentinc.atlassian.net/wiki/spaces/CNKAF/pages/1219931556/SBK+Remove+Broker+Mega-Pager, this patch introduces an internal Kafka RPC client for SBK.

It is similar to Kafka's own ReplicaFetcherBlockingSend class, in that it can send requests to one broker only, at most one request concurrently and blocks until the response is received. This is the behavior that SBK will need for calling the InitiateShutdown request as part of the remove broker operation.

There are also some small changes to EmbeddedKafka to make it friendlier to double-shutdown (one from the test, another from the test teardown)

This patch adds a test dependency of ce-sbk->ce-broker-plugins, for the EmbeddedKafkaCluster integration helpers. Future ce-sbk integration tests should be written with those, as like the rest of ce-kafka's non-core code