KCORE-426: Server should not resolve DNS on establishment of SASL connection (#1681)
We are now using the same code for SASL and SSL, the latter was updated
to have the correct behavior a while back via 17b2bde4be6.

Ported the unit tests from 17b2bde4be6 to SaslAuthenticatorTest, two of
them failed without this change. Also updated some unit tests to use
localhost instead of 127.0.0.1 so that they pass after these changes.

Finally, included a few minor clean-ups.

Credit to @progocz for finding the issue in CCloud and filing the ticket. This
was also independently found by @gardnervickers and @lbradstreet.

Reviewers: David Arthur <mumrah@gmail.com>