MINOR: Fix scala 2.13 compilation errors (#1688)
scala.collection.IterableView is scala.collection.View now and nither of
them exist in other version (one is in 2.12 and another one is in 2.13).
This change takes the type declaration out so that Scala infers it based
on its version. The type was added for code clarity, but can't have that
now.

In addition JavaConverters is deprecated, so needed to add a
SuppressWarning.

compilation with `-PscalaVersion=2.13` passes now. It used to fail