CNKAF-697 (part1): Increase min broker quota constants (#1701)
Per-broker quota for a tenant is set to 32KB/s if the broker has not tenant's leaders.
During a roll, when a broker restarted, initially it has no leaders and will get 32KB/s quota.
When a client start send requests to the new broker after preferred leader change, there
maybe a small window when the quota is not updated. This causes large throttle times.

This PR changes the constants on a broker to be 10MB/s so it works OK for multi-tenant
clusters (there maybe a small window where a tenant may get more quota than it should),
and works for enterprise clusters. We will create a follow up PR to fix related broker configs ("confluent.quota.tenant.broker.min.consumer.rate",
"confluent.quota.tenant.broker.min.producer.rate") because looks like they are not declared
in KafkaConfig and then set those configs in cc-spec-kafka to larger values for enterprise
(single-tenant) clusters.

Reviewers: Jun Rao <junrao@gmail.com>, Ismael Juma <ismael@juma.me.uk>