KCORE-416: Fix flaky Observer test (#1651)
* KCORE-416: Fix flaky Observer test

The testRackAwareWithConstraintReplicaAssignmentWithStartIndex was
testing against head of a hashmap. A HashMap isn't supposed to be
ordered and thus was resulting in random failures. This change instead
checks for the containment in the HashMap.

Test passes locally. I wasn't able to reproduce locally, but as the code
and assertion is changed now, flakiness should be gone.