[CNKAF-608] During SBK shutdown allow for CruiseControl termination (#1621)
In DataBalanceEngine::shutdown, wait for up to 15s to let ongoing executions start terminating.

This opens shutdown() up to possible InterruptedExceptions, which are propagated up to the DataBalanceManager and captured there. This should be OK since this is only called during actual shutdown anyways.
Lower the execution progress check interval to 7s from 10s, to make it much more likely that an execution check (and execution cancellation). will occur during broker shutdown.