CNKAF-701: Refactor SampleStore topic existence check (#1711)
* CNKAF-701: Refactor SampleStore topic existence check

This change moves topic check and subsequent creation if it is mising
from SampleStore to a new TopicUtils class. This allows us to use the
same code for other topics that need same functionality (like topic to
store state of api calls). The move also parameterized the code so that
its easier to test for existence of different topics with different
properties.

KafkaSampleStore test continues to pass and added a new test class to
test the new code that was introduced.