KSTORAGE-627: TierPartitionForceRestore, TierPartitionState hotswap and tier task recovery (#1471)
In this PR we implement handling for TierPartitionForceRestore metadata as described in https://confluentinc.atlassian.net/wiki/spaces/KSTORAGE/pages/1153008788/One+Pager+Log+Divergence+State+Injection+Mechanism.

TierPartitionForceRestore requires a few different elements to be implemented safely.

A restoreOffsetAndEpoch, and startOffset have been added to the FileTierPartitionState header. The restoreOffsetAndEpoch is the offset and epoch of the message that triggered the last restore. All segment metadata and the TierPartitionForceRestore metadata now includes a stateOffsetAndEpoch which is used to check the validity of that metadata to ensure that it was generated for the current state and not an older TierPartitionState for a previous restore.

Upon reading TierForcePartitionRestore metadata a partition in ERROR state will be restored to the blessed version of the state uploaded to object storage. Whenever state is restored, it is stamped with the epoch and offset of the tier topic record containing the restored state.

The last piece to ensuring a successful recovery is that TierPartitionState fencing/ERROR status is now no longer a permanent TierTask failure, and is now a failure with retry. Upon retry, the TierTask will now check for a return to non-ERROR status, and upon detecting one will ensure that the TierPartitionState is setup with the correct epoch and will restart any archiving and deletion activity.

Tooling and system tests to inject the new state and trigger a recovery are left for a follow up PR, currently tracked at lbradstreet#3.

Tests:
- FileTierPartitionStateTest: adds test for state hotswap behavior
- TierPartitionStateConcurrencyTest: hotswap concurrency test
- TierTopicManagerTest: TierPartitionForceRestore handling in TierTopicConsumer
- TierMessageFormatterTest: formatting of TierPartitionForceRestore metadata
- ArchiveTaskTest: awaken tasks in error state
- DeletionTaskTest: awaken tasks in error state