CNKAF-649: Introduce reservations in SBK's Executor (#1730)
This patch revises the access strategy used for the Executor. The remove broker feature will need to reserve the Executor prior to making use of the class (during broker shutdown and plan computation) in order to ensure that self-healing (or anything else) doesn't trigger another proposal execution under its feet.

Further, because we have the notion of a "critical operation" that overrides others (as per the UX doc), it's useful to have a straightforward way of cancelling ongoing operations/locks and taking over.