KC-382: Allow alter cluster configs for subset of configs in CCloud (#1802)
This is enabled via `confluent.alter.cluster.configs.enable`. Initially, cc-spec-kafka will
set this to true for whitelisted (via Launch Darkly) dedicated clusters. Later,
we may enable this for every dedicated clusters.

This PR also includes:
1. Make `AutoCreateTopicsEnableProp` dynamic.
2. Updated AlterConfigsPolicy to potentially allow cluster config updates
for `ssl.cipher.suites`, `auto.create.topics.enable`, `num.partitions` and
`log.retention.ms`. Note that we only allow cluster config updates via
the external listener. Broker specific updates can only be done via
the internal listener.
3. `AlterConfigsPolicy` also restricts the valid values to safe values.
For cipher suites, we only allow 5 TLS 1.2 ciphers. We will extend
these with the TLS 1.3 ciphers when we add support for that (TLS 1.3
specifies 5 ciphers). The allowed range for `num.partitions` and
`log.retention.ms` is a bit arbitrary and errs on the side of caution.
We can relax them later.
4. Unit tests for the updated AlterConfigsPolicy.
5. Integration tests that verify the behavior of the policy in combination
with the interceptor.
6. Renamed TenantIsolationTest to MultiTenantKafkaIntegrationTest.
7. A few code clean-ups, mostly in tests.

What needs to be done in future PRs:
1. Integrate auto topic creation with the create topic policy and return
a non retriable error that clients expect from MetadataRequest. I
suggest using `TopicAuthorizationFailed`.
2. The interceptor should prefix ``ssl.cipher.suites` with the external
listener name on the way in and remove the prefix on the way out.
The `AlterConfigsPolicy` should be adjusted to take this into account.
3. Adjust the code so that `ssl.cipher.suites` can be updated cluster wide.
This is currently not allowed in AK and some changes are needed to fix
that. Once we do this, we need to update `MultiTenantKafkaIntegrationTest`
to test it (remove `Ignore` from the test that has it) and add cipher suite
updates to `testAlterClusterConfigsWhenConfigEnabled`.
4. Test performance of `TLS_ECDHE_RSA_CHACHA20_POLY1305_SHA256`
as we have never tested it before. It's a secure cipher that tends to be
used by mobile devices. We don't seem to have any CCloud users using
it.

Reviewers: Agam Brahma <agam.brahma+git@gmail.com>