[CNKAF-741] Handle duplicate metrics in the DataBalancerMetricsRegistry (#1789)
The DataBalancerMetricsRegistry now allows duplicate metric names to be registered. Update
some of its unit tests to allow this, and add test cases for when a duplicate name is attempted to be registered for a long-lived metric even though a short-lived one already exists.

The DataBalancerMetricsRegistry will still throw if users attempt to  register short- and long-lived metrics with the same name.