CPKAFKA-4744; Fix test failure due to concurrent cleaning (#1829)
We are seeing this test failure fairly regularly. The cause of the failure is a race condition updating the last modified time of the segment. This is because the call to `reconfigureMaxSegmentDeletedPerRun`, which is used to initialize the value of this configuration, also starts the cleaner. If we get unlucky, then the cleaner may be trying to update the last modified time the same time as the test case. The patch fixes the problem by initializing the configuration value directly when the cleaner is constructed so that we can modify the segment last modified times before the cleaner is started. I haven't been able to reproduce the failure after fixing this.