KSTORAGE-730: Allow max.partition.fetch.bytes to be configurable for tier fetches (#1754)
Tier fetches are a bit different from regular consumer fetches in that we include data for a single tiered partition in the response, up to the configured partition.max.fetch.bytes. In some cases where the cluster is overloaded, we have seen this cause issues for the consumer where the lag keeps increasing if the number of requests being processed on the broker is low enough.

This PR introduces a dynamic configuration on the broker `confluent.tier.max.partition.fetch.bytes.override` that allows overriding this maximum fetch bytes per partition, so we are able to match the overall throughput to the case where tiering were disabled. We compute the maximum possible bytes that could have been returned without tiered storage and fetch as many bytes as that and the minimum of the configured override.