CNKAF-649 - Compose the broker removal operation into multiple interruptible tasks (#1812)
This patch breaks down the 4-step broker removal operation into chained, composable futures that get returned to the DataBalanceEngine so that it has the ability to cancel the operation.

Whenever an interrupted exception is raised, we immediately complete the future exceptionally and return it. 

This patch moves all the classes into their own package to make things more organized. It also moves BrokerRemovalCallback into com.linkedin.cruisecontrol.brokerremoval as most of its usages are inside that parent package (com.linkedin.cruisecontrol)