MINOR: Remove the notion of a TrainingTask (#1843)
This patch removes the notion of a TrainingTask in CruiseControl. A TrainingTask was a task which ran metric sampling in a loop for a certain time period. It was part of the original CruiseControl code and was called via an explicit /train HTTP request

Since we are not making use of this feature, and are unlikely to in the near-future, it makes sense to remove it and reduce confusion when reading the code.