CC-8138: Add Connect feature to inject Confluent license properties into connector properties before startup (#1840)
Implemented the feature that allows Confluent license properties to be defined in the Connect worker configuration, so that when Confluent licensed connectors are deployed the user no longer has to include the Confluent license properties. Instead, the worker dynamically injects its Confluent license properties into the connector configuration for Confluent-licensed connectors (which normally include the license properties in their ConfigDef objects) and filters the validation output such that the user doesn't even know the connector needs these properties.

These Confluent license properties are injected just before the Confluent-licensed connector is started, and the connector configuration persisted by the user is never modified. 

If a persisted connector configuration already has Confluent license properties, they are used as is and the worker does not inject any of its Confluent license properties. This allows for easy migration/upgrade, and for users to override their own Confluent license properties when needed.

This feature is enabled by default, but can be disabled with `confluent.license.inject.into.connectors=false`.

The configurations of connectors that are not Confluent-licensed will not be modified.

New unit and integration tests were added. The integration tests check the behavior when the feature is enabled and when it is disabled, with connectors that are Confluent-licensed and those that are not.