CNKAF-760: Integrate api state persistence with remove broker (#1841)
* CNKAF-760: Integrate api state persistence with remove broker

This change swaps the in-memory api state persistence used by remove
broker api with the topic based persistence added as part of CNKAF-752.
In addition to that this change also refactors some of the repetetive
code in the ConfluentDataBalanceEngine. The change also adds
BrokerAddStatus that can be used by AddBroker api to persist its state.

All the unit tests have been modified appropriately to work with topic
based persistence.