KCORE-54: Controller should accumulate metadata into fewer requests to mitigate handling delays (#1685)
This PR introduced metadata batching in the controller. The overall idea is to accumulate metadata changes and to turn them into requests when the sender thread is ready to sent the next request instead of queuing up requests to be sent.

The PR replaces the traditional control requests by control metadata batches that implements the ControlMetadataBatch trait. Three batch types are defines: LeaderAndIsrBatch, StopReplicaBatch, and UpdateMetadataBatch. Unlike their requests counter parts, batches can be merged and invalidated.

When a batch is queued up in the ControllerChannelManager, the internal queue takes advantage of the ControlMetadataBatch to merge it in an already existing batch of same type and to invalidate stale metadata updates of other types. If there is not batch of the same type already pending, the batch is inserted at the end of the queue.