Add schemas topic label to SchemaRegistryProvider (#1897)
In order to properly identify the SR instance from telemetry data, we use a combination of Kafka cluster ID and backend schemas topic. This PR adds picks up the topic name from the SR configuration and adds it as a label to the metrics.

Integration testing occurs in the SR telemetry reporter.