CNKAF-575: Remove IOtec:ZkClient dependency. (#1878)
* CNKAF-575: Remove IOtec:ZkClient dependency.

This commit replaces IOTec:ZKClient with KafkaZKClient in
BrokerFailureDetector.

As part of this change, BrokerFailureDetector now runs in its own
thread and responds to broker change notifications from zookeeper
concurrently without the need for polling. Previously, the IOtec
ZKClient implementation processed notifications and called the
listener on a separate thread.

This change removes the need for any synchronization within
BrokerFailureDetector as all operations are done on a single thread.
AnomalyDetector and ZookeeperClient threads, post events to the
BrokerFailureDetector eventQueue to prompt failure detection.
The BrokerFailureDetector thread continuously drains the events and
detects/reports broker failures.
The integration tests have been modified to use the new mechanism.