CNKAF-649: Handle broker restart during broker removal (#1854)
This PR adds handling for the scenario where a broker is restarted while a removal operation is being performed on it. A special cancelled status is set, along with a new public API exception BrokerRemovalCanceledException. Both get shown to the user when the describe API is called.

This patch adds one new broker removal state - BROKER_SHUTDOWN_CANCELED. It's called when the broker to be removed is restarted.

The cancellation logic lives in KafkaDataBalanceManager and ConfluentDataBalanceEngine.
KafkaDataBalanceManager is responsible for setting the appropriate state via the stateTracker (which persists to memory and would persist to disk) and then makes the DataBalanceEngine cancel the futures. In order for DataBalanceEngine to be able to cancel the operation, we extend KafkaCruiseControl to return the ExecutionProposalRunnable future.