Persist the broker removal state and handle failover (#1875)
This patch introduces support for SBK failover of remove broker API operations.
To do this, we:

Now persist the state of the broker removal operation to disk, rather than the statuses
On SBK enabling (dynamic config or Controller startup), we re-submit broker removal operations that have not reached a terminal state
Throw exceptions when the state tracker is cancelled and callers attempt to register new events - this ensures that background threads running the removal operations eventually get fenced when the state tracker is cancelled. Note we expect them to stop when we cancel them, but this gives us a stronger guarantee.
Split executable phases into a new BrokerRemovalRestartablePhase class, which allows us to skip phases up to the one the operation last processed pre-failover.
Store StateTrackers in-memory in ConfluentDataBalanceEngine, and introduce a new BrokerRemovalTerminationListener called when the broker removal operation reaches a terminal state, at which point we clean the state tracker in-memory.
Renamed internal representations of the broker removal operation's status. We now have - BrokerRemovalStateRecordProto (auto-generated proto object), BrokerRemovalStateRecord (ce-sbk representation of a broker removal), BrokerRemovalDescriptionInternal (core package representation of a broker removal) and BrokerRemovalDescription (client-side and public API representation of a broker removal)