CNKAF-847: Complete removal callback when there are no proposals to execute (#1977)
CNKAF-773 introduced a regression where we wouldn't ever complete the broker removal operation's status on an empty plan execution for broker removal. This commit fixes the issue and ensures broker removals are always completed

This wasn't caught by the integration test because of CNKAF-838 which will be fixed with a separate patch.