Allowing user to disable tiered storage safely - Phase 1 (#1775)
With this change, users will be able to disable tier storage for a given topic using incremental alter configs. This change was not allowed earlier.
This is phase 1 of the project where we do not bring the tiered data to local disks. Tiered data
stays at object store where it can be fetched from and it gets deleted per the retention configuration. No new data is archived.

Co-authored-by: Matthew Wong <mattwong949@gmail.com>