CIAM-94 Decouple scope path keys and Role scope type (#1954)
* CIAM-94 Decouple scope path keys and Role scope type

Previously we used "scopeType" to define whether a role was a
"Cluster" role or a "Resource" role. Resource roles applied
only to specific resources within a cluster, and Cluster
roles gave permissions on all resources with the cluster.

With Confluent Cloud RBAC, we need to be able to assign roles
at Organization and Environment scopes. This caused the
"scopeType" to be overloaded. We were using it to indicate
both what "level" of scope a policy applied to *and* whether
it should be bound with resources.

This commit breaks apart those two concepts. "bindingScope"
tells what level of scope a policy applies to, and
"requiresResource" tells whether a binding must be made
with resources.