CNKAF-773: Do not throw an exception when the broker is not present in the cluster model (#2001)
It is possible to have cases where we try to remove a broker that's not present in the cluster model - this can happen when the broker is shut down (leaves metadata) and has no replicas on it. This is a rare edge case because we have validation at the API layer against this, but if said validation has passed we should still have handling for this scenario.

This patch makes it such that we complete the broker removal successfully, log a warning on such an occasion during broker removal (be it API-driven or self-healing driven) and swallow the exception.