CNKAF-893: Update SBK to work with secure Zookeeper (#2041)
This change updates SBK code to work when Zookeeper is set with
both TLS and SASL based authentication. It does this by getting rid
of simple zkClient usages and for more complicated case by passing
appropriate zk configuration.

Tested on a secure cluster to make sure SBK starts and can serve
api requests. System test will be added for this as a separate PR.