CNKAF-849: Sanitize log messages when checking for startup conditions (#2094)
When we check for topics needed by SBK to become available, we may
encounter transient errors. These errors are currently reported using throwing/catching
exception. Logging those create noise in the log. Just printing the error message and
retrying is better.

This change updates the error handling to log message at info level and skip
throwing exception. If the unlikely case an exception is thrown because of a
real non-transient error, the exception is continued to be logged at ERROR level.