METRICS-1835 Shade confluent-serializers dependencies in confluent-metrics (#2111)
Some projects directly depend on  confluent-serializers at compile time, but also
need confluent-metrics at runtime. This ensures the confluent-serializers classes
included in the confluent-metrics fat jar do not clash with confluent-serializers
classes independently pulled into the classpath.

Additionally, since we shade some protobuf and guava dependencies for confluent-serializers
and confluent-licensing as part of ce-broker-plugins, we also need to do the same in ce-sbk to
avoid ending up with the two versions of the confluent-serializers classes with protobuf shaded
differently on the broker classpath.