[CNKAF-885, CNKAF-946, CNKAF-947] Set SBK Default Throttle, handle absent brokers when setting throttle. (#2119)
* [CNKAF-885] Ensure that the default ConfluentConfigs throttle value goes through to CruiseControl.
* [CNKAF-946] Treat a throttle value of NO_THROTTLE as disabled (previously treated as invalid).
* [CNKAF-947]: Allow setting replication throttle when some brokers are offline.
Due to the way the ReplicationThrottleHelper worked, it would not be able to set replication throttles for brokers (X,Y,Z) if any subset of the brokers were offline. This is because we describe the configs for each broker in order to know whether they have statically-set replication throttles, which allows SBK to conditionally override them depending on a configurable variable.

This patch changes the overrides and config describes to be best effort - that is, if the config describe for a given broker fails, we continue with setting replication throttles for it (via ZK) and the rest of the brokers. Any static throttles on said offline broker would always get overridden, as we do not have the broker configs handy to know whether we are overriding or not. In practice, this should not be an issue since the broker is offline regardless and the configs get cleaned up afterwards

* Fix tests: integration tests, and unit tests due to different API calls (fix up mocks). Ensure that throttle is set for RemoveBrokerTest to avoid a reprise of CNKAF-947.

* Rename DEFAULT_REPLICATION_THROTTLE_CONFIG property, include doc for disabling throttling.

Co-authored-by: Stanislav Kozlovski <stanislav_kozlovski@outlook.com>