KC-532: Support observers in replica assignment tool (#1889)
Instead of using a plain sequence to describe replica assignments, we created a case class that includes both replica assignments and observer assignments. This way observer reassignments can now be read into the partition reassignment program. We also added an error message for the "generate" option if any topics in the reassignment JSON have topic placement constraints. Use confluent-rebalancer or SBK instead to generate observer reassignments.

In the integration tests we had to do something differently than what was done before since we are assigning observers through topic placement constraints. When constraints are used, replica and observer assignments are not deterministic. First we create a topic, then use the client to query for a description of the topic. This gave us the initial replica and observer assignments. We then generate a reassignment JSON that includes replicas and observers that are not in the initial assignment. This dynamically generated reassignment is used to test the program.